requirements_deploy.txt
django_rocket_engine